+++ src/api/auth/auth.controller.ts
@@ -1,10 +1,9 @@
-import { Body, Controller, Post, Res, Session, UseGuards } from '@nestjs/common';
+import { Body, Controller, Get, Post, Res, Session, UseGuards } from '@nestjs/common';
 import { Response } from 'express';
 import { SessionContainer } from 'supertokens-node/recipe/session';
-import { UserLoginDto } from '../users/users.dto';
+import { EmailDto, PasswordDto, UserLoginDto } from '../users/users.dto';
 import { AuthGuard } from './auth.guard';
 import { AuthService } from './auth.service';
-import { STSession } from './supertokens/supertokens.types';
 
 @Controller('auth')
 export class AuthController {
@@ -15,9 +14,20 @@ export class AuthController {
 		return this.authService.login(userLoginDto, resp);
 	}
 
-	@Post('logout')
+	@Get('logout')
 	@UseGuards(AuthGuard)
 	async logout(@Session() session: SessionContainer) {
 		return this.authService.logout(session);
+	}
+
+	@Post('password/reset')
+	async resetPassword(@Body() emailDto: EmailDto) {
+		return this.authService.sendPasswordResetTokenEmail(emailDto.email);
+	}
+
+	@Post('password/reset/token')
+	async performPasswordReset(@Body() passwordResetDto: PasswordDto) {
+		const { token, password } = passwordResetDto;
+		return this.authService.performPasswordReset(token, password);
 	}
 }