+++ src/api/auth/config.interface.ts
@@ -2,4 +2,4 @@ import { AppInfo } from 'supertokens-node/lib/build/types';
 
 export const ConfigInjectionToken = 'ConfigInjectionToken';
 
-export type AuthModuleConfig = { appInfo: AppInfo; connectionURI: string; apiKey?: string };
+export type AuthModuleConfig = { appInfo: AppInfo; connectionURI: string; apiKey?: string; apiBasePath?: string };