+++ src/db/prisma/migrations/20220619202234_password_resets/migration.sql
@@ -1,0 +1,13 @@
+/*
+  Warnings:
+
+  - A unique constraint covering the columns `[passwordResetToken]` on the table `user` will be added. If there are existing duplicate values, this will fail.
+
+*/
+-- AlterTable
+ALTER TABLE "user" ADD COLUMN     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+ADD COLUMN     "passwordResetToken" TEXT,
+ALTER COLUMN "passwordHash" DROP NOT NULL;
+
+-- CreateIndex
+CREATE UNIQUE INDEX "user_passwordResetToken_key" ON "user"("passwordResetToken");