+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2022-06-19
+- Added password reset flow
+- Integrated SuperTokens' email verificaion recipe
+
 ## 2022-06-17
 - Changed project to use pnpm
 - Replaced ESLint/Prettier with XO